+++ src/Adapter/HashedName.php
@@ -115,9 +115,9 @@ final class HashedName implements Adapter
         $extension = \pathinfo($name->toString(), \PATHINFO_EXTENSION);
         $hash = Str::of(\sha1(\pathinfo($name->toString(), \PATHINFO_BASENAME)));
 
-        $first = new Name($hash->substring(0, 2)->toString());
-        $second = new Name($hash->substring(2, 2)->toString());
-        $remaining = new Name($hash->substring(4)->toString().($extension ? '.'.$extension : ''));
+        $first = Name::of($hash->substring(0, 2)->toString());
+        $second = Name::of($hash->substring(2, 2)->toString());
+        $remaining = Name::of($hash->substring(4)->toString().($extension ? '.'.$extension : ''));
 
         return [$first, $second, $remaining];
     }